[feat] (front) 공구글의 참여 취소 버튼 추가, 공구글에 참여시, 모달 추가 및 조건 추가

+++ BackEnd_Modeul/src/main/java/com/modeul/web/service/ParticipationService.java
@@ -19,5 +19,5 @@ public interface ParticipationService {
         Long getMemberCountBystuffId(Long stuffId);
     
         /* 멤버별 참여 목록 수 집계 */
-        Long getStuffCountBymemberId(Long memberId);
+        int getStuffCountBymemberId(Long memberId);
 }
